refactor: Restructure wasmtime_embedder module
