+++ rs/embedders/src/wasmtime_embedder/signal_handler.rs
